requirements.txt
opencv-python==4.5.3
torch==1.8.2
torchvision==0.9.2
pytesseract==0.3.8
Pillow==8.3.2
numpy==1.21.2
